+++ resources/views/components/line-annotated.blade.php
@@ -1,0 +1,13 @@
+@props([
+    'series' => [],
+    'categories' => [],
+    'annotations' => [],
+    'smooth' => true,
+    'legend' => true,
+    'average' => false,
+])
+
+<x-chart::base component="line-annotated" :option="\WireCharts\Support\Option::lineAnnotated([
+    'series' => $series, 'categories' => $categories, 'annotations' => $annotations,
+    'smooth' => $smooth, 'legend' => $legend, 'average' => $average,
+])" {{ $attributes }} />